removed not needed `syncBy` and moved `syncWith` to `WithJob` interface
